+++ hydra-node/src/Hydra/Chain/Offline.hs
@@ -6,12 +6,11 @@ import Cardano.Api.Genesis (shelleyGenesisDefaults)
 import Cardano.Api.Ledger qualified as Ledger
 import Cardano.Ledger.Coin qualified as L
 import Cardano.Ledger.Shelley.Genesis qualified as Shelley
-import Cardano.Ledger.Slot (unSlotNo)
 import Cardano.Slotting.Time (SystemStart (SystemStart), mkSlotLength)
 import Control.Monad.Class.MonadAsync (link)
 import Data.Aeson qualified as Aeson
 import Data.Aeson.Types qualified as Aeson
-import Hydra.Cardano.Api (GenesisParameters (..), NetworkMagic (..), ShelleyEra, ShelleyGenesis (..), Tx, fromShelleyNetwork)
+import Hydra.Cardano.Api (ChainPoint (..), GenesisParameters (..), NetworkMagic (..), ShelleyEra, ShelleyGenesis (..), Tx, fromShelleyNetwork)
 import Hydra.Chain (
   Chain (..),
   ChainComponent,
@@ -19,17 +18,19 @@ import Hydra.Chain (
   ChainStateHistory,
   OnChainTx (..),
   PostTxError (..),
-  chainSlot,
+  chainPoint,
   chainTime,
   initHistory,
  )
-import Hydra.Chain.ChainState (ChainSlot (ChainSlot))
 import Hydra.Chain.Direct.State (initialChainState)
 import Hydra.Ledger.Cardano.Time (slotNoFromUTCTime, slotNoToUTCTime)
 import Hydra.Node.Util (checkNonADAAssetsUTxO)
 import Hydra.Options (OfflineChainConfig (..), defaultContestationPeriod)
 import Hydra.Tx (HeadId (..), HeadParameters (..), HeadSeed (..), Party, Snapshot (..), getSnapshot)
 import Hydra.Utils (readJsonFileThrow)
+import Test.Gen.Cardano.Api.Typed (genBlockHeaderHash)
+import Test.QuickCheck (generate)
+import Test.QuickCheck.Hedgehog (hedgehog)
 
 -- Upstreamed in cardano-api 10.18
 fromShelleyGenesis :: Shelley.ShelleyGenesis -> GenesisParameters ShelleyEra
@@ -200,10 +201,11 @@ tickForever genesis callback = do
     let timeToSleepUntil = slotNoToUTCTime systemStart slotLength upcomingSlot
     sleepDelay <- diffUTCTime timeToSleepUntil <$> getCurrentTime
     threadDelay $ realToFrac sleepDelay
+    blockHash <- generate (hedgehog genBlockHeaderHash)
     callback $
       Tick
         { chainTime = timeToSleepUntil
-        , chainSlot = ChainSlot . fromIntegral $ unSlotNo upcomingSlot
+        , chainPoint = ChainPoint upcomingSlot blockHash
         }
   systemStart = SystemStart protocolParamSystemStart