+++ hydra-tx/src/Hydra/Chain/ChainState.hs
@@ -2,6 +2,7 @@
 
 module Hydra.Chain.ChainState where
 
+import Hydra.Cardano.Api (ChainPoint (..), SlotNo (..))
 import Hydra.Prelude
 
 import Hydra.Tx (IsTx (..))
@@ -27,6 +28,15 @@ class
   -- XXX: Why is this not always UTxOType?
   type ChainStateType tx = c | c -> tx
 
-  -- | Get the chain slot for a chain state. NOTE: For any sequence of 'a'
-  -- encountered, we assume monotonically increasing slots.
-  chainStateSlot :: ChainStateType tx -> ChainSlot
+  -- | Get the chain point for a chain state.
+  chainStatePoint :: ChainStateType tx -> ChainPoint
+
+-- | Get the chain slot for a chain point.
+chainPointSlot :: ChainPoint -> ChainSlot
+chainPointSlot = \case
+  ChainPointAtGenesis -> ChainSlot 0
+  ChainPoint SlotNo{unSlotNo} _ -> ChainSlot $ fromIntegral unSlotNo
+
+-- encountered, we assume monotonically increasing slots.
+chainStateSlot :: IsChainState tx => ChainStateType tx -> ChainSlot
+chainStateSlot = chainPointSlot . chainStatePoint