+++ CHANGELOG.md
@@ -18,6 +18,11 @@ changes.
   - See [Issue #2286](https://github.com/cardano-scaling/hydra/issues/2286) and [PR #2290](https://github.com/cardano-scaling/hydra/pull/2290).
 - Fix the `cabalOnly` development environment to allow for `cabal build hydra-node` without `haskell.nix`.
 
+- **BREAKING** Resume node from the latest observed tick when `--start-chain-from` is not set.
+  - `TickObserved` event schema has changed: the `chainSlot` field has been replaced with `chainPoint`
+  - `Greetings` message now also contains a new field `atChainPoint` that represents the last known chain point observed by the chain backend.
+  - See [Issue #2206](https://github.com/cardano-scaling/hydra/issues/2206) and [PR #2407](https://github.com/cardano-scaling/hydra/pull/2407)
+
 ## [1.2.0] - 2025.11.28
 
 - There is a new `SafeClose` client command which prevents closing the Head in case there are non-ADA assets in the confirmed snapshot UTxO [#2330](https://github.com/cardano-scaling/hydra/issues/2330).